refactor: sperate each codemod to have its own spec file
